backend/requirements.txt
flask==2.0.1
firebase-admin==5.3.0
gunicorn==20.1.0
python-dotenv==0.19.2
Pillow==9.3.0
flask-cors==3.0.10 